chore: apply the mechanical analyzer fixes and stop analysing build output
`dart fix` applied for six codes only, all of which are removals rather than
rewrites: unused_import, unnecessary_import, unnecessary_cast, unnecessary_this,
unnecessary_new and unnecessary_const. 475 fixes across 344 files, clearing all
49 unused imports the earlier migration steps left behind.

The remaining lints are stylistic (widget key constructors, doc comment slashes,
SizedBox for whitespace, child property ordering) and are deliberately left
alone so the revival diff stays reviewable.

analysis_options.yaml now excludes build/. The app displays its own source, so
every demo page is declared as an asset and `flutter build` copies it into the
bundle. The analyzer was picking up those copies and reporting 190 phantom
errors about relative imports that only resolve from lib/.

+++ lib/src/pages/quotes/quotes1.dart
@@ -40,7 +40,7 @@ class QuotesOnePage extends StatelessWidget {
               tween: Tween<double>(begin: -1, end: 0),
               builder: (context, state, child) {
                 return FractionalTranslation(
-                    translation: Offset(state.value as double, 0), child: child);
+                    translation: Offset(state.value, 0), child: child);
               },
               child: Text(
                 "Anyone who has never made a mistake has never tried anything new",
@@ -55,7 +55,7 @@ class QuotesOnePage extends StatelessWidget {
               tween: Tween<double>(begin: 1, end: 0),
               builder: (context, state, child) {
                 return FractionalTranslation(
-                  translation: Offset(state.value as double, 0),
+                  translation: Offset(state.value, 0),
                   child: child,
                 );
               },
